# Rewrite the GAP to Julia conversion