--- a/src/gap_to_julia.jl
+++ b/src/gap_to_julia.jl
@@ -453,6 +453,8 @@ It just delegates to `gap_to_julia`.
 Its purpose is to turn the `recursive` argument into a keyword argument,
 which is easier in Julia than in GAP.
 """
+function _gap_to_julia end
+
 _gap_to_julia(x::Obj) = gap_to_julia(x)
 _gap_to_julia(::Type{T}, x::Obj) where {T} = gap_to_julia(T, x)
 _gap_to_julia(::Type{T}, x::Obj, recursive::Bool) where {T} =
@@ -465,4 +467,3 @@ _gap_to_julia(x::GapObj, recursive::Bool) = gap_to_julia(x; recursive)
 _gap_to_julia(x::Bool, recursive::Bool) = x
 _gap_to_julia(x::Int, recursive::Bool) = x
 _gap_to_julia(x::FFE, recursive::Bool) = x
-