--- a/src/gap_to_julia.jl
+++ b/src/gap_to_julia.jl
@@ -23,11 +23,22 @@ Base.showerror(io::IO, e::ConversionError) =
     print(io, "failed to convert $(typeof(e.obj)) to $(e.jl_type):\n $(e.obj)")
 
 """
-    RecDict_j
+    RecDict_j = IdDict{Tuple{Any, Type}, Any}
 
 An internal type of GAP.jl used for tracking conversion results in `gap_to_julia`.
+The value stored at the key `(obj, T)` is the result
+of the GAP to Julia conversion of `obj` that has type `T`.
+Note that several Julia types can occur for the same GAP object.
+
+Lookups for the key `(obj, T)` in an `IdDict` are successful if the conversion
+result of an object identical to `obj` with target type `T` has been stored
+in the dictionary.
+
+Note that comparing two `GapObj`s with `===` yields the same result as
+comparing them with `GAP.Globals.IsIdenticalObj`
+because `GapObj` is a mutable type.
 """
-const RecDict_j = IdDict{Any, IdDict{Type, Any}}
+const RecDict_j = IdDict{Tuple{Any, Type}, Any}
 
 const JuliaCacheDict = Union{Nothing,RecDict_j}
 
@@ -96,7 +107,7 @@ end
 ## Now come conversions that support recursion.
 
 # helper functions for recursion in conversion from GAP to Julia
-function recursion_info_g(::Type{T}, obj, recursive::Bool, recursion_dict::JuliaCacheDict) where {T}
+function recursion_info_j(::Type{T}, obj, recursive::Bool, recursion_dict::JuliaCacheDict) where {T}
     if recursive && recursion_dict === nothing
         return RecDict_j()
     else
@@ -106,10 +117,8 @@ end
 
 function handle_recursion_g(obj, ::Type{T}, ret_val, rec::Bool, rec_dict::JuliaCacheDict) where T
     if rec_dict !== nothing
-      if !haskey(rec_dict, obj)
-        rec_dict[obj] = IdDict{Type, Any}()
-      end
-      rec_dict[obj][T] = ret_val
+      # We assume that `obj` is not yet cached.
+      rec_dict[(obj, T)] = ret_val
     end
     return rec ? rec_dict : nothing
 end
@@ -123,8 +132,8 @@ function gap_to_julia_internal(
     ::Val{recursive},
 ) where {T, recursive}
 
-    recursive && recursion_dict !== nothing && haskey(recursion_dict, obj) && haskey(recursion_dict[obj], TT) && return recursion_dict[obj][TT]
-    rec_dict = recursion_info_g(TT, obj, recursive, recursion_dict)
+    recursive && recursion_dict !== nothing && haskey(recursion_dict, (obj, TT)) && return recursion_dict[(obj, TT)]
+    rec_dict = recursion_info_j(TT, obj, recursive, recursion_dict)
 
     if Wrappers.IsList(obj)
         islist = true
@@ -165,8 +174,8 @@ function gap_to_julia_internal(
     ::Val{recursive},
 ) where {T, recursive}
 
-    recursive && recursion_dict !== nothing && haskey(recursion_dict, obj) && haskey(recursion_dict[obj], TT) && return recursion_dict[obj][TT]
-    rec_dict = recursion_info_g(TT, obj, recursive, recursion_dict)
+    recursive && recursion_dict !== nothing && haskey(recursion_dict, (obj, TT)) && return recursion_dict[(obj, TT)]
+    rec_dict = recursion_info_j(TT, obj, recursive, recursion_dict)
 
     if Wrappers.IsMatrixObj(obj)
         nrows = Wrappers.NumberRows(obj)::Int
@@ -206,8 +215,8 @@ function gap_to_julia_internal(
     ::Val{recursive},
 ) where {T, recursive}
 
-    recursive && recursion_dict !== nothing && haskey(recursion_dict, obj) && haskey(recursion_dict[obj], TT) && return recursion_dict[obj][TT]
-    rec_dict = recursion_info_g(TT, obj, recursive, recursion_dict)
+    recursive && recursion_dict !== nothing && haskey(recursion_dict, (obj, TT)) && return recursion_dict[(obj, TT)]
+    rec_dict = recursion_info_j(TT, obj, recursive, recursion_dict)
 
     if Wrappers.IsCollection(obj)
         newobj = Wrappers.AsSet(obj)
@@ -238,14 +247,14 @@ end
 ## thus we have to convert at least also the next layer,
 ## even if `recursive == false` holds.
 function gap_to_julia_internal(
-    ::Type{T},
+    TT::Type{T},
     obj::GapObj,
     recursion_dict::JuliaCacheDict,
     ::Val{recursive},
 ) where {T<:Tuple, recursive}
 
-    recursive && recursion_dict !== nothing && haskey(recursion_dict, obj) && haskey(recursion_dict[obj], T) && return recursion_dict[obj][T]
-    rec_dict = recursion_info_g(T, obj, recursive, recursion_dict)
+    recursive && recursion_dict !== nothing && haskey(recursion_dict, (obj, TT)) && return recursion_dict[(obj, TT)]
+    rec_dict = recursion_info_j(TT, obj, recursive, recursion_dict)
 
     !Wrappers.IsList(obj) && throw(ConversionError(obj, T))
     parameters = T.parameters
@@ -270,8 +279,8 @@ function gap_to_julia_internal(
     ::Val{recursive},
 ) where {T, recursive}
 
-    recursive && recursion_dict !== nothing && haskey(recursion_dict, obj) && haskey(recursion_dict[obj], TT) && return recursion_dict[obj][TT]
-    rec_dict = recursion_info_g(TT, obj, recursive, recursion_dict)
+    recursive && recursion_dict !== nothing && haskey(recursion_dict, (obj, TT)) && return recursion_dict[(obj, TT)]
+    rec_dict = recursion_info_j(TT, obj, recursive, recursion_dict)
 
     !Wrappers.IsRecord(obj) && throw(ConversionError(obj, TT))
     dict = TT()