- Introduce an internal function `gap_to_julia_internal` for which the (perhaps recursive) methods get installed that do the work.
- Delegate from `gap_to_julia` to `gap_to_julia_internal`.
- Introduce a new object type for caching objects and their Julia types, in order to keep the relation between identical objects. Create these objects only if they are needed.
- Rename some internal types, in order to distinguish between the two conversion directions. We need different objects for bookkeeping in recursive situations.

Due to the changes, there are no "back delegations" anymore. The behaviour changed slightly:

- One test has a different result now, I think the new result is better.
- The method for converting GAP objects to Julia `Set`s was wrong w.r.t. the identity objects in recursive situations; a new test shows the difference.

The changes are breaking in the sense that new methods for GAP to Julia conversion shall now get installed for `gap_to_julia_internal` not for `gap_to_julia`.

addresses #1089