@ThomasBreuer 2025-08-20
That's strange:
I am sure there was a comment by @fingolfin on the conversion to `Set{T}` in `src/gap_to_julia.jl`,
and that I responded to this comment. How did the two comments disappear?