@fingolfin 2025-08-20
The comments are still there, just marked as "resolved" and hence "collapsed"