@lgoettgens 2025-08-21
> Unfortunately I am still seeing a performance regression in some cases. E.g. using the example from issue #736, in Julia 1.10.10:

I just tried to reproduce you benchmarks on my machine. With julia 1.10.10, I am getting similar regressions as you (for the same lines).
However, when using 1.11.6 for the benchmark, none of the cases worsens with this PR. Actually, a bunch of them (both for `Vector{GapObj}` and `Vector{Int}`) need less allocations and time with this PR (compared to master). (One has to note though that comparing 1.10.10 to 1.11.6 won't work that well as there are a lot of changes in julia that heavily influence the results)