@fingolfin 2025-08-21
Some more timings on my mac, this time with Julia 1.12.0-rc1 (in light of what @lgoettgens reported):


master:
```julia-repl
julia> using GAP, BenchmarkTools

julia> L = GAP.evalstr("List([3..6],SymmetricGroup)")
GAP: [ Sym( [ 1 .. 3 ] ), Sym( [ 1 .. 4 ] ), Sym( [ 1 .. 5 ] ), Sym( [ 1 .. 6 ] ) ]

julia> @btime collect($L);
  1.913 μs (23 allocations: 880 bytes)

julia> @btime [x for x in $L];
  1.875 μs (24 allocations: 896 bytes)

julia> @btime Vector{GapObj}($L);
  178.540 ns (8 allocations: 480 bytes)

julia> v = GapObj([1,2,3,4]);

julia> @btime Vector{Int}($v);
  98.608 ns (8 allocations: 480 bytes)

julia> @btime Vector{Int}($v; recursive=false);
  108.539 ns (8 allocations: 480 bytes)

julia> function f(x::GapObj) v=Vector{Int}(undef,length(x)); for i in 1:length(x) v[i]=x[i] end ;  v end;

julia> @btime f($v);
  59.978 ns (2 allocations: 96 bytes)
```

This PR:
```julia-repl
julia> using GAP, BenchmarkTools

julia> L = GAP.evalstr("List([3..6],SymmetricGroup)")
GAP: [ Sym( [ 1 .. 3 ] ), Sym( [ 1 .. 4 ] ), Sym( [ 1 .. 5 ] ), Sym( [ 1 .. 6 ] ) ]

julia> @btime collect($L);
  1.833 μs (23 allocations: 880 bytes)

julia> @btime [x for x in $L];
  1.796 μs (24 allocations: 896 bytes)

julia> @btime Vector{GapObj}($L);
  112.928 ns (5 allocations: 448 bytes)

julia> v = GapObj([1,2,3,4]);

julia> @btime Vector{Int}($v);
  137.534 ns (5 allocations: 448 bytes)

julia> @btime Vector{Int}($v; recursive=false);
  95.608 ns (2 allocations: 96 bytes)

julia> function f(x::GapObj) v=Vector{Int}(undef,length(x)); for i in 1:length(x) v[i]=x[i] end ;  v end;

julia> @btime f($v);
  100.854 ns (2 allocations: 96 bytes)
```


So: `Vector{GapObj}` got faster but `Vector{Int}` slower