@fingolfin 2025-08-25
So I redid my `Vector{Int}` measurement, where it looked as if it got slower with this PR. New results:

`master` + Julia 1.10.10
```julia-repl
julia> v = GapObj([1,2,3,4]); @btime Vector{Int}($v);
  210.998 ns (7 allocations: 496 bytes)
```

`master` + Julia 1.11.6
```julia-repl
julia> v = GapObj([1,2,3,4]); @btime Vector{Int}($v);
  171.167 ns (8 allocations: 480 bytes)
```

`master` + Julia 1.12.0-rc1
```julia-repl
julia> v = GapObj([1,2,3,4]); @btime Vector{Int}($v);
  82.345 ns (8 allocations: 480 bytes)
```

This PR + Julia 1.10.10
```julia-repl
julia> v = GapObj([1,2,3,4]); @btime Vector{Int}($v);
  524.521 ns (8 allocations: 624 bytes)
```

This PR + Julia 1.11.6
```julia-repl
julia> v = GapObj([1,2,3,4]); @btime Vector{Int}($v);
  150.633 ns (8 allocations: 480 bytes)
```

This PR + Julia 1.12.0-rc1
```julia-repl
julia> v = GapObj([1,2,3,4]); @btime Vector{Int}($v);
  72.842 ns (5 allocations: 448 bytes)
```

So... it's complicated... Fastest is this PR and Julia 1.12, slowest ist also this PR but with Julia 1.10. Ugh.

But this branch is also faster than master when used in Julia 1.11, so I'd say: let's merge this, and I'll open a bunch of issues so we can improve more later.